+++ prisma/migrations/20260816102935_create_route_likes/migration.sql
@@ -1,7 +1,7 @@
 -- CreateTable
 CREATE TABLE `likes` (
-    `user_id` BIGINT NOT NULL,
-    `route_id` BIGINT NOT NULL,
+    `user_id` INTEGER NOT NULL,
+    `route_id` INTEGER NOT NULL,
     `created_at` DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
 
     INDEX `likes_route_id_idx`(`route_id`),
